replace the murmur3 impl w/ a more portable one

+++ test/lib_murmur.c
@@ -5,7 +5,7 @@
 int main(int argc, char **argv)
 {
   char hex[9];
-  uint32_t *str = (uint32_t*)"foo bar"; // TODO fix this!
+  uint8_t *str = (uint8_t*)"foo bar"; // TODO fix this!
   size_t len = strlen((char*)str);
 
   fail_unless(murmur4(str,len) == 171088983l);
